% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cobweb.R
\name{geom_cobweb}
\alias{geom_cobweb}
\title{Cobweb plot}
\usage{
geom_cobweb(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = TRUE,
  theme = list(panel.background = element_rect(fill = "white", colour =
  "white")), grid_lty = "dashed", grid_lwd = 1, grid_alpha = 0.75,
  grid_fontsize = 12, grid_Nticks = 4, grid_ticks = NULL,
  grid_labels = NULL, grid_label_position = 1,
  grid_label_hjust = 0.5, grid_label_vjust = -0.5, ...)
}
\arguments{
\item{mapping}{(from ggplot call) A mapping of ggplot variables. geom_cobweb
expects an AUCs variable.}

\item{data}{(from ggplot call) The data frame supplied to the}

\item{stat}{(from ggplot call) ggplot identity stat.}

\item{position}{(from ggplot call) ggplot identity position.}

\item{na.rm}{(from ggplot call) remove NAs, default is FALSE.}

\item{show.legend}{Whether to show the legend. Default is TRUE.}

\item{theme}{Default theme settings for correct visualization.}

\item{grid_lty}{The linetype of the background grid. Deafult is 'dashed'.}

\item{grid_lwd}{The linewidth of the background grid. Deafult is 1.}

\item{grid_alpha}{The alpha of the background grid. Deafult is 0.75.}

\item{grid_fontsize}{The fontsize of the grid labels. Deafult is 12.}

\item{grid_Nticks}{The number of gridlines (makes use of
\code{\link[base]{pretty}} so number is not a hard threshold). Default is
4.}

\item{grid_ticks}{vector with values of gridlines to be drawn (unlike
grid_Nticks, this is exact).}

\item{grid_labels}{The labels to use for the gridlines.}

\item{grid_label_position}{The position alonng the cobweb where to plot the
grid labels. (numeric value)}

\item{grid_label_hjust}{The hjust of grid_labels}

\item{grid_label_vjust}{The vjust of grid_labels}

\item{...}{Extra aesthetics parameters, see Aesthetics section.}
}
\description{
The cobweb geom is an addition to the \pkg{ggplot2} package for plotting
cobweb or spider plots.
}
\section{Aesthetics}{
 \code{geom_cobweb} understands the following aesthetics
  (required aesthetics are in bold): 
  \itemize{ 
  \item \strong{\code{AUCs}} 
  \item \code{alpha} 
  \item \code{colour} 
  \item \code{group} 
  \item \code{linetype} 
  \item \code{size} }
}

\examples{
library(ggplot2)
n_comparison <- 3
AUC.df <- data.frame(type = c(rep("ROC",6),rep("Random",6)),  
                    AUCs = c(0.34, 1.00, 0.56, 0.40, 0.37, 0.45, rep(1/n_comparison, 6)))

ggplot(AUC.df, aes(AUCs = AUCs,  colour = type)) + 
    geom_cobweb() + 
    theme_cobweb() +
    ggtitle("Testen") +
    coord_equal()
 
 
 
ggplot(AUC.df, aes(AUCs = AUCs,  colour = type)) + 
    geom_cobweb(grid_Nticks = 2,
                show.legend = TRUE,
                grid_label_position = 2) + 
    theme_cobweb() +
    ggtitle("Testen") +
    coord_equal()
 
}
