% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cobweb_bg_proto.R
\docType{data}
\name{geom_cobweb_bg_proto}
\alias{geom_cobweb_bg_proto}
\title{Helper function for geom_cobweb}
\format{An object of class \code{geom_cobweb_bg_proto} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.}
\usage{
geom_cobweb_bg_proto
}
\description{
This ggproto function draws the background stuff.
}
\keyword{datasets}
