% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rad2polar2xy.R
\name{rad2polar2xy}
\alias{rad2polar2xy}
\title{Helper function for polar plotting}
\usage{
rad2polar2xy(x)
}
\arguments{
\item{x}{The values to convert to evenly spaced polar coordinates, then to x,y coordinates.}
}
\description{
\code{rad2polar2xy} Converts single values to polar coordinates, with the points evenly spaced out.
These points are then transformed to cartesian coordinates for plotting with geom_cobweb.
}
