% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cobweb.R
\name{theme_cobweb}
\alias{theme_cobweb}
\title{Theme for cobweb visualization}
\usage{
theme_cobweb(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
\code{theme_cobweb} Provides the theme to the geom_cobweb ggplot2 function. It is advised
to also add coord_equal() to the plot to achieve a cobweb with of equal lengths.
}
\examples{
library(ggplot2)
n_comparison = 3
AUC.df = data.frame(type = c(rep("ROC",6),rep("Random",6)),  
                    AUCs = c(0.34, 1.00, 0.56, 0.40, 0.37, 0.45, rep(1/n_comparison, 6)))
ggplot(AUC.df, aes(AUCs = AUCs,  colour = type)) +
    geom_cobweb() + 
    theme_cobweb() +
    ggtitle("Test plot") +
    coord_equal()

}
